mlp works and matches pytorch backward + forward

+++ numpy_sequential/mlp.py
@@ -6,11 +6,11 @@ import numpy_sequential as nseq
 class MLP:
     """Simple Multi-Layered Perceptron with two layers."""
 
-    def __init__(self, d_model: int, d_hidden: int, rng):
+    def __init__(self, d_model: int, d_hidden: int, rng, dtype=np.float32):
         self.layers = [
-            nseq.Linear(d_model=d_model, d_hidden=d_hidden, rng=rng),
+            nseq.Linear(d_model=d_model, d_hidden=d_hidden, rng=rng, dtype=dtype),
             nseq.ReLU(),
-            nseq.Linear(d_model=d_hidden, d_hidden=d_model, rng=rng),
+            nseq.Linear(d_model=d_hidden, d_hidden=d_model, rng=rng, dtype=dtype),
         ]
 
     def forward(self, inputs: np.ndarray) -> np.ndarray:
